requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
google-cloud-aiplatform==1.42.0
google-auth==2.27.0
python-docx==1.1.0
reportlab==4.0.7
